fix(datepicker): circular dependency in common types
