+++ src/DatePicker/index.ts
@@ -1,2 +1,3 @@
 export { default as DatePicker } from './DatePicker';
-export type { DatePickerProps, RangeSelectionValue, CalendarPlacement, DateFormat } from './DatePicker';
+export type { DatePickerProps }  from './DatePicker'
+export type {RangeSelectionValue, CalendarPlacement, DateFormat } from './types';
