+++ src/DatePicker/types.ts
@@ -1,0 +1,7 @@
+export interface RangeSelectionValue {
+  start: Date | undefined;
+  end: Date | undefined;
+}
+
+export type CalendarPlacement = 'up' | 'down';
+export type DateFormat = 'MM/DD/YYYY' | 'DD/MM/YYYY' | 'YYYY/MM/DD';
