+++ src/utils/getTotalDaysInMonth.ts
@@ -1,0 +1,3 @@
+export const getTotalDaysInMonth = (date: Date) => {
+    return new Date(date.getFullYear(), date.getMonth() + 1, 0).getDate();
+  };
